refactor: move RecommendedTracks type into track type file
