--- src/components/Main/Recommendation.tsx
@@ -1,9 +1,0 @@
-import { Recommendations } from '@/models/recommendation';
-
-export default function Recommendation({
-  recommendations,
-}: {
-  recommendations: Recommendations | null;
-}) {
-  return <>{JSON.stringify(recommendations)}</>;
-}
