+++ src/components/Main/Recommendations.tsx
@@ -1,0 +1,9 @@
+import { RecommendedTracks } from '@/models/track';
+
+export default function Recommendations({
+  recommendations,
+}: {
+  recommendations: RecommendedTracks | null;
+}) {
+  return <>{JSON.stringify(recommendations)}</>;
+}
